rubocop auto changes

+++ app/controllers/concerns/deposit_data.rb
@@ -6,37 +6,32 @@ module DepositData
   require 'zip'
 
   included do
-
   end
 
-  def new_deposit(dataset_id,aip_id)
+  def new_deposit(dataset_id, aip_id)
     @dir_dataset = ENV['TRANSFER_LOCATION'] + '/' + dataset_id
     @dir_aip = @dir_dataset + '/' + aip_id + '/'
     make_data_directories
   end
 
   def make_data_directories
-    unless Dir.exists? (@dir_dataset)
-      FileUtils.mkdir(@dir_dataset)
-    end
-    unless Dir.exists? (@dir_aip)
-      FileUtils.mkdir(@dir_aip)
-    end
+    FileUtils.mkdir(@dir_dataset) unless Dir.exist? @dir_dataset
+    FileUtils.mkdir(@dir_aip) unless Dir.exist? @dir_aip
   end
 
-  # given an array of files on the user's client machine, upload them, unzip them if they're zipped, and store them in the AIP folder 
+  # given an array of files on the user's client machine, upload them, unzip them if they're zipped, and store them in the AIP folder
   def deposit_files_from_client(files)
     # handle each of the uploaded files
     files.each do |f|
       # get the uploaded filename (including its path in the case of directory upload)
       h = HttpHeaders.new(f.headers)
-      uploaded_filename = h.content_disposition.match(/filename=(\"?)(.+)\1/)[2] 
+      uploaded_filename = h.content_disposition.match(/filename=(\"?)(.+)\1/)[2]
       # if it's a .zip file, extract its contents to @dir_aip
-      if File.extname(uploaded_filename) == '.zip' then
+      if File.extname(uploaded_filename) == '.zip'
         Zip::File.open(f.tempfile) do |zip_file|
           zip_file.each do |entry|
             # extract everything except mac osx guff
-            unless entry.name.include?("__MACOSX") then
+            unless entry.name.include?('__MACOSX')
               entry.extract(File.join(@dir_aip, entry.name))
             end
           end
@@ -49,22 +44,22 @@ module DepositData
         # create any directories in target_dir that don't already exist
         FileUtils.mkdir_p(target_dir)
         # move this uploaded file into place (into target_dir)
-        FileUtils.chmod 0644, f.tempfile
+        FileUtils.chmod 0o644, f.tempfile
         FileUtils.mv(f.tempfile, target_file)
       end
     end
   end
 
   # given a hash of Google Drive files selected by the user, download them to the AIP folder
-#  def deposit_files_from_cloud(files)
-#    retriever = BrowseEverything::Retriever.new
-#    files.each do |index, file|
-#      target_file = File.join(@dir_aip, file['file_name'])
-#      retriever.download(file, target_file) do |filename, retrieved, total|
-#        # could potentially output download progess here
-#      end
-#    end
-#  end
+  #  def deposit_files_from_cloud(files)
+  #    retriever = BrowseEverything::Retriever.new
+  #    files.each do |index, file|
+  #      target_file = File.join(@dir_aip, file['file_name'])
+  #      retriever.download(file, target_file) do |filename, retrieved, total|
+  #        # could potentially output download progess here
+  #      end
+  #    end
+  #  end
 
   def deposit_files_from_cloud(files, paths, mime_types)
     # initialise the google api
@@ -74,36 +69,33 @@ module DepositData
       # download the file from Google
       f = get_file_from_google(service, file, mime_type)
       # if the file was a google document it will have been exported to a specific format and may require an extra file extension
-      if google_docs_mimetypes.has_key?(mime_type)
-        path = path + google_docs_mimetypes[mime_type]["export_extension"] unless path.ends_with?(google_docs_mimetypes[mime_type]["export_extension"])
+      if google_docs_mimetypes.key?(mime_type)
+        path += google_docs_mimetypes[mime_type]['export_extension'] unless path.ends_with?(google_docs_mimetypes[mime_type]['export_extension'])
       end
       # work out where this file should be uploaded to
       target_file = File.join(@dir_aip, path)
       target_dir = File.dirname(target_file)
       FileUtils.mkdir_p(target_dir)
-      File.open(target_file, "wb") do |output|
+      File.open(target_file, 'wb') do |output|
         output.write f.string
-      end 
+      end
     end
   end
 
   # delete all files deposited in the AIP - this will be called to clean things up if there was a problem during file upload
   def delete_deposited_files
-    if @dir_aip
-      FileUtils.rm_rf(@dir_aip)
-    end
+    FileUtils.rm_rf(@dir_aip) if @dir_aip
   end
 
   def add_metadata(metadata)
     require 'json'
-    json = JSON.generate(JSON.parse metadata.gsub('=>', ':'))
+    json = JSON.generate(JSON.parse(metadata.gsub('=>', ':')))
     File.write(@dir_aip + 'metadata.json', json)
   end
 
   def add_submission_documentation
-    unless Dir.exists? (dir + 'submissionDocumentation')
+    unless Dir.exist? (dir + 'submissionDocumentation')
       FileUtils.mkdir(dir + 'submissionDocumentation')
     end
   end
-
 end
